Split out Step 3 into its own view, refactor, and add JS for max age
Also, adjust wonky comma placement
