+++ retirement_api/templates/claiming.html
@@ -392,19 +392,19 @@
             <div id="age-selector-response">
               <hr>
               <div class="next-step-description next-steps_max">
-                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value">\, </span> {% trans "which is your maximum Social Security benefit claiming age." %}</h3>
+                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value"></span>, {% trans "which is your maximum Social Security benefit claiming age." %}</h3>
                 <p>{% trans "This means you will get your maximum possible monthly benefit! Here are steps to help you in the next few years." %}</p>
               </div>
               <div class="next-step-description next-steps_over">
-                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value">\, </span> {% trans "which is later than your Social Security full retirement claiming age." %}</h3>
+                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value"></span>, {% trans "which is later than your Social Security full retirement claiming age." %}</h3>
                 <p>{% trans "This means you will get an increase in your monthly benefits!  Here are steps to help you in the next few years." %}</p>
               </div>
               <div class="next-step-description next-steps_equal">
-                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value">\, </span> {% trans "which is your Social Security full retirement claiming age." %}</h3>
+                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value"></span>, {% trans "which is your Social Security full retirement claiming age." %}</h3>
                 <p>{% trans "That means you’ll receive your full benefit with no reductions! Here are steps to help you in the next few years." %}</p>
               </div>
               <div class="next-step-description next-steps_under">
-                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value">\, </span> {% trans "which is earlier than your Social Security full retirement claiming age." %}</h3>
+                <h3>{% trans "You’ve chosen age" %} <span class="age-response-value"></span>, {% trans "which is earlier than your Social Security full retirement claiming age." %}</h3>
                 <p>{% trans "At this age you will receive less than your full benefit amount. Here are steps to help you in the next few years." %}</p>
               </div>
               <ol class="next-steps content-l">
