+++ src/js/index.js
@@ -2,12 +2,14 @@
 
 var graphView = require( './views/graph-view' );
 var questionsView = require( './views/questions-view' );
+var nextStepsView = require( './views/next-steps-view' );
 var tooltipsView = require( './views/tooltips-view' );
 
 var app = {
   init: function() {
     graphView.init();
     questionsView.init();
+    nextStepsView.init();
     tooltipsView.init();
   }
 };
